+++ core/core.js
@@ -4,6 +4,7 @@
 // in the bootstrap.
 export { render } from "../src/render.js";
 export { buildItemData, filenameFields, zoteroSelectURI, ensureZoteroLink } from "../src/item-data.js";
+export { resolveNoteFilename } from "../src/filename.js";
 export { mergeNote, updateNoteAnnotations } from "../src/merge.js";
 export { renderAnnotationsSection, mapZoteroAnnotation } from "../src/annotations.js";
 export { syncBlocks, makeBlock, parseBlocks, migrateLegacyAnnotations } from "../src/blocks.js";
@@ -12,7 +13,7 @@ export { parseTemplateFile, templateKind, templateUserOwnedKeys } from "../src/t
 export { parseManifest, hasManifest, applyManifest, setManifestEntry, removeManifestEntry, buildManifestFromScaffold, writeManifest, MANIFEST_KEY, getTagField, setTagField, getAttachmentFolder, setAttachmentFolder } from "../src/manifest.js";
 export { frontmatterList, cleanTag, tagSyncPlan } from "../src/tagsync.js";
 export { findMarkerRanges, rangeRevealed } from "../src/markers.js";
-export { findFrontmatterRange, findHeadingRanges, findLinkRanges, findEmphasisRanges } from "../src/preview.js";
+export { findFrontmatterRange, findHeadingRanges, findLinkRanges, findEmphasisRanges, findImageEmbedRanges } from "../src/preview.js";
 export { COLOR_NAMES } from "../src/colors.js";
 export { buildCrossrefURL, pickBestMatch, normalizeTitle, titleSimilarity, normalizeDOI, crossrefYear, extractYear } from "../src/crossref.js";
 export { splitPath, vaultName, vaultRelative, buildObsidianUri, obsidianConfigPath, parseObsidianVaults, sanitizeFilename, isUnder } from "../src/paths.js";